Add files via upload
This is an app that only chooses person from once contact number and then been added the blood group manually that supplied by that contact person. This app will show which persons are already gave blood within 3 months or recently or more than 3 months and produce green, red, blue colored signal to confirm to the user that to whom he can call for his/her necessary blood (grouped).
Its a cloud based app but need more to work or implementation  . In future this app will be more extended and more flexible to general.